refactor(node-toolbar): replace `ViewportTransform` import with `Viewport`
